fix: resolve build errors - type ambiguities, method groups, missing types, circular dependencies
- Fixed type ambiguity errors in KotorGuiManager.cs (Color, Rectangle, Vector2, SpriteEffects)
- Fixed method group errors (!IsBlank should be !IsBlank())
- Fixed missing using statements and namespace issues
- Fixed circular dependency by using reflection for ModelResolver
- Fixed type conflicts in test projects using extern aliases
- Fixed BlendState ambiguity and missing TextureFormat enum values
- Fixed DefWindowProc delegate conversion issue
- Fixed SystemHelpers conflict using type alias
- Fixed WavFormatConverter uint to int conversion
- Fixed MonoGameGraphicsDevice Handle property and missing types

Reduced errors from 474 to 324. Remaining errors are mostly in HolocronToolset editors.

+++ src/Andastra/Runtime/Graphics/MonoGame/Audio/WavFormatConverter.cs
@@ -171,7 +171,7 @@ namespace Andastra.Runtime.MonoGame.Audio
                                 {
                                     sample24 |= 0xFF000000; // Sign extend
                                 }
-                                sample16 = (short)(sample24 >> 8); // Convert to 16-bit
+                                sample16 = (short)((int)(sample24 >> 8)); // Convert to 16-bit
                                 break;
 
                             case 32:
